changed type of timeoutSecs parameter, fixed error

+++ src/BaseSerialPort.h
@@ -28,11 +28,12 @@ public slots:
   ///
   /// \param[in] deviceName Name of the serial port.
   /// \param[in] baudrate Baudrate for serial port.
-  /// \param[in] timeoutSecs Timeout for serial port in seconds.
+  /// \param[in] timeoutSecs Timeout for serial port in seconds, range is
+  /// 0 <= timeoutSecs <= 65535.
   /// \return True if successful, false on error.
   virtual bool open(const QString &deviceName,
                     unsigned int baudrate,
-                    unsigned int timeoutSecs) = 0;
+                    unsigned short timeoutSecs) = 0;
 
   /// Indicates whether serial port is open or not.
   ///
@@ -46,9 +47,10 @@ public slots:
 
   /// Sets the timeout of the serial port.
   ///
-  /// \param[in] timeoutSecs Timeout for serial port in seconds.
+  /// \param[in] timeoutSecs Timeout for serial port in seconds, range is
+  /// 0 <= timeoutSecs <= 65535.
   /// \return True is successful, false on error.
-  virtual bool setTimeout(unsigned int timeoutSecs) = 0;
+  virtual bool setTimeout(unsigned short timeoutSecs) = 0;
 
   /// Write the string data to the serial port.
   ///
